msys clang: OKed

+++ src/3rdParty/libE57Format/src/SourceDestBufferImpl.cpp
@@ -487,7 +487,7 @@ int64_t SourceDestBufferImpl::getNextInt64( double scale, double offset )
          throw E57_EXCEPTION2( E57_ERROR_INTERNAL, "pathName=" + pathName_ );
    }
    /// Make sure that value is representable in an int64_t
-   if ( doubleRawValue < E57_INT64_MIN || E57_INT64_MAX < doubleRawValue )
+   if ( doubleRawValue < (double)E57_INT64_MIN || (double)E57_INT64_MAX < doubleRawValue )
    {
       throw E57_EXCEPTION2( E57_ERROR_SCALED_VALUE_NOT_REPRESENTABLE,
                             "pathName=" + pathName_ + " value=" + toString( doubleRawValue ) );
